requirements.txt
streamlit
pandas
numpy
requests
matplotlib
altair
plotly
Pillow
PyPDF2
